Add files via upload
5- Write a program that uses 6 threads to compute the sum of an array of 3000 integers.
The program should divide the set to 6 parts and give the 1st thread the first 500 integers , the 2nd
thread the next 500 integers and so on. The 6 sums from these parts should then be added to give a
total sum.
The program should display this total sum.
The program should read the dataset of a 3000 integers from a file. This file name should be given in
the command line (argc, argv).