Add files via upload
6- Write a program that does as follows : Create an array of 100 integers (You are free to
create this array however you like, hard code it in the code, read from a file , take it from the command
line ). Create two threads that read this array. The 1st Thread should compute the sum of the 100
integers. The 2nd Thread should compute the Average of the 100 integers. Display the Sum and
Average in the end.